losenconfirm på deleteblog

+++ app/Controller/Blog.php
@@ -92,7 +92,8 @@ class Blog extends Controller
     }
 
      public function settings($args = [])
-    {
+    {   
+        $confirmPassword = (isset($_POST['confirmpassword'])) ? trim($_POST['confirmpassword']) : '';
         $blogname = $this->blogName;
         $blog_id = $this->model('blog')->getBlogId($blogname);
         if(!$this->userModel ->isLoggedIn())
@@ -104,11 +105,17 @@ class Blog extends Controller
         if ($auth != 7) {
           Redirect::to('/'.$blogname);
         }
-        if(isset($_POST['delete'])){
+        if (empty($confirmPassword)) {
+            UserError::add(Lang::FORM_CONFIRMATION_PASSWORD_SENT_NO);
+        } 
+        if (!password_verify($confirmPassword, $this->userModel->get($this->userModel->getLoggedInUserId())->password)) {
+            UserError::add(Lang::FORM_PASSWORD_ORIGINAL_INVALID);
+        }
+        if(isset($_POST['delete']) && !empty($confirmPassword) && password_verify($confirmPassword, $this->userModel->get($this->userModel->getLoggedInUserId())->password) == true){
           $blog_id = $_POST['delete'];
           $bloggen = $this->model('Blog')->deleteBlog($blog_id);
           Redirect::to('/dashboard');
-        }
+        } 
 
         $search = [];
